Completed Polygon class

+++ src/matrix4.cpp
@@ -1,7 +1,7 @@
 #include "matrix4.h"
 
 Matrix4::Matrix4():
-    elements{new std::array<float, 16>{{0}}}
+    elements{new std::array<double, 16>{{0}}}
 {
 
 }
